fix: bug fix modify instorage with unit is wrong.

+++ bean/src/main/java/com/cotte/estate/bean/pojo/ao/storage/OutStorageAo.java
@@ -53,7 +53,7 @@ public class OutStorageAo {
 
     private String count;
 
-    private Integer bunchCount;
+    private BigDecimal bunchCount;
 
     private String bake;
